requirements.txt
numpy
pillow
click
Flask
Flask-HTTPAuth
Flask-SQLAlchemy
itsdangerous
Jinja2
MarkupSafe
passlib
SQLAlchemy
Werkzeug
